refactor: 메인 페이지 카테고리 버튼 폴더를 main 디렉토리 안으로 이동
